requirements.txt
rpaframework
pywin32



